requirements.txt
asgiref==3.8.1
contourpy==1.2.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
distlib==0.3.9
Django>=4.2,<5.0
filelock==3.18.0
fonttools==4.60.1
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
numpy==2.2.4
packaging==25.0
pandas==2.2.3
phonenumbers==8.13.55
pillow==11.3.0
platformdirs==4.3.7
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.35.4
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
